+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/ui/NumberOfShowsStory.kt
@@ -91,7 +91,8 @@ internal fun NumberOfShowsStory(
                     story.showCount,
                     story.epsiodeCount,
                 ),
-                disableScale = true,
+                disableAutoScale = true,
+                fontScale = measurements.smallDeviceFactor,
                 color = colorResource(UR.color.coolgrey_90),
                 modifier = Modifier.padding(horizontal = 24.dp),
             )
@@ -101,7 +102,7 @@ internal fun NumberOfShowsStory(
             TextP40(
                 text = stringResource(R.string.end_of_year_story_listened_to_numbers_subtitle),
                 fontSize = 15.sp,
-                disableScale = true,
+                disableAutoScale = true,
                 color = colorResource(UR.color.coolgrey_90),
                 modifier = Modifier.padding(horizontal = 24.dp),
             )
@@ -139,7 +140,7 @@ private fun PodcastCoverCarousel(
 @Preview(device = Devices.PortraitRegular)
 @Composable
 private fun NumberOfShowsPreview() {
-    PreviewBox { measurements ->
+    PreviewBox(currentPage = 1) { measurements ->
         NumberOfShowsStory(
             story = Story.NumberOfShows(
                 showCount = 20,