+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/viewmodel/PlayerViewModel.kt
@@ -18,6 +18,7 @@ import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
 import au.com.shiftyjelly.pocketcasts.models.to.Chapter
 import au.com.shiftyjelly.pocketcasts.models.to.Chapters
 import au.com.shiftyjelly.pocketcasts.models.to.PlaybackEffects
+import au.com.shiftyjelly.pocketcasts.models.type.EpisodeStatusEnum
 import au.com.shiftyjelly.pocketcasts.player.R
 import au.com.shiftyjelly.pocketcasts.player.view.UpNextPlaying
 import au.com.shiftyjelly.pocketcasts.player.view.bookmark.BookmarkArguments
@@ -27,6 +28,8 @@ import au.com.shiftyjelly.pocketcasts.preferences.model.ArtworkConfiguration
 import au.com.shiftyjelly.pocketcasts.preferences.model.ShelfItem
 import au.com.shiftyjelly.pocketcasts.repositories.bookmark.BookmarkManager
 import au.com.shiftyjelly.pocketcasts.repositories.di.ApplicationScope
+import au.com.shiftyjelly.pocketcasts.repositories.download.DownloadHelper
+import au.com.shiftyjelly.pocketcasts.repositories.download.DownloadManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackState
 import au.com.shiftyjelly.pocketcasts.repositories.playback.SleepTimer
@@ -79,6 +82,7 @@ class PlayerViewModel @Inject constructor(
     private val userEpisodeManager: UserEpisodeManager,
     private val podcastManager: PodcastManager,
     private val bookmarkManager: BookmarkManager,
+    private val downloadManager: DownloadManager,
     private val sleepTimer: SleepTimer,
     private val settings: Settings,
     private val theme: Theme,
@@ -107,6 +111,7 @@ class PlayerViewModel @Inject constructor(
         val skipBackwardInSecs: Int = 30,
         val isSleepRunning: Boolean = false,
         val isEffectsOn: Boolean = false,
+        val adjustRemainingTimeDuration: Boolean = false,
         val playbackEffects: PlaybackEffects = PlaybackEffects(),
         val isBuffering: Boolean = false,
         val bufferedUpToMs: Int = 0,
@@ -123,7 +128,11 @@ class PlayerViewModel @Inject constructor(
         val isChaptersPresent: Boolean = !chapters.isEmpty
         val chapter: Chapter? = chapters.getChapter(positionMs.milliseconds)
         val chapterProgress: Float = chapter?.calculateProgress(positionMs.milliseconds) ?: 0f
-        val chapterTimeRemaining: String = chapter?.remainingTime(positionMs.milliseconds, playbackEffects.playbackSpeed) ?: ""
+        val chapterTimeRemaining: String = chapter?.remainingTime(
+            playbackPosition = positionMs.milliseconds,
+            playbackSpeed = playbackEffects.playbackSpeed,
+            adjustRemainingTimeDuration = adjustRemainingTimeDuration,
+        ) ?: ""
         val chapterSummary: String = chapters.getChapterSummary(positionMs.milliseconds)
         val isFirstChapter: Boolean = chapters.isFirstChapter(positionMs.milliseconds)
         val isLastChapter: Boolean = chapters.isLastChapter(positionMs.milliseconds)
@@ -179,6 +188,7 @@ class PlayerViewModel @Inject constructor(
         settings.skipForwardInSecs.flow.asObservable(coroutineContext),
         upNextExpandedObservable,
         chaptersExpandedObservable,
+        settings.useRealTimeForPlaybackRemaingTime.flow.asObservable(coroutineContext),
         settings.artworkConfiguration.flow.asObservable(coroutineContext),
         this::mergeListData,
     )
@@ -340,7 +350,16 @@ class PlayerViewModel @Inject constructor(
             }
     }
 
-    private fun mergeListData(upNextState: UpNextQueue.State, playbackState: PlaybackState, skipBackwardInSecs: Int, skipForwardInSecs: Int, upNextExpanded: Boolean, chaptersExpanded: Boolean, artworkConfiguration: ArtworkConfiguration): ListData {
+    private fun mergeListData(
+        upNextState: UpNextQueue.State,
+        playbackState: PlaybackState,
+        skipBackwardInSecs: Int,
+        skipForwardInSecs: Int,
+        upNextExpanded: Boolean,
+        chaptersExpanded: Boolean,
+        adjustRemainingTimeDuration: Boolean,
+        artworkConfiguration: ArtworkConfiguration,
+    ): ListData {
         val podcast: Podcast? = (upNextState as? UpNextQueue.State.Loaded)?.podcast
         val episode = (upNextState as? UpNextQueue.State.Loaded)?.episode
 
@@ -377,6 +396,7 @@ class PlayerViewModel @Inject constructor(
                 isSleepRunning = playbackState.isSleepTimerRunning,
                 isEffectsOn = !effects.usingDefaultValues,
                 playbackEffects = effects,
+                adjustRemainingTimeDuration = adjustRemainingTimeDuration,
                 isBuffering = playbackState.isBuffering,
                 bufferedUpToMs = playbackState.bufferedMs,
                 theme = theme.activeTheme,
@@ -513,6 +533,22 @@ class PlayerViewModel @Inject constructor(
                 tintColor = tintColor,
             )
             onSuccess(arguments)
+        }
+    }
+
+    fun handleDownloadClickFromPlaybackActions(onDeleteStart: () -> Unit, onDownloadStart: () -> Unit) {
+        val episode = playbackManager.upNextQueue.currentEpisode ?: return
+
+        if (episode.episodeStatus != EpisodeStatusEnum.NOT_DOWNLOADED) {
+            onDeleteStart.invoke()
+            launch {
+                episodeManager.deleteEpisodeFile(episode, playbackManager, disableAutoDownload = false, removeFromUpNext = episode.episodeStatus == EpisodeStatusEnum.DOWNLOADED)
+            }
+        } else {
+            onDownloadStart.invoke()
+            launch {
+                DownloadHelper.manuallyDownloadEpisodeNow(episode, "Player shelf", downloadManager, episodeManager, source = source)
+            }
         }
     }